major revamp of the extension - v.4.0.0
